Extracted form data parser into separate file
